start refactor of torch nn classes
